anlyzer_backend/pyproject.toml
[project]
name = "anlyzer-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "fastapi>=0.129.0",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.1.0",
    "uvicorn>=0.40.0",
]
